requirements.txt
feedparser>=6.0.12
requests>=2.31.0
openai>=1.51.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
pydantic>=2.5.0
pydantic-settings>=2.11.0
psutil>=5.9.0
python-dateutil>=2.8.0
redis>=5.0.0
aioredis>=2.0.0